Support for multiple detection policies
Factor out the current inception_and_threshold policy from
detect_fire.py and add "always" and "never" test policies.
The already bitrotted minusMinutes code took a bit more damage
in this factorization, but it will be revived soon as a new
detection policy.
